ps renderers now use gvio

+++ tclpkg/gv/gv_php_init.c
@@ -20,12 +20,12 @@
 #include "gvcjob.h"
 #include "gvcint.h"
 
-static size_t gv_string_writer (GVJ_t *job, const char *s, int len)
+static size_t gv_string_writer (GVJ_t *job, const char *s, size_t len)
 {
     return PHPWRITE(s, len);
 }
 
-static size_t gv_channel_writer (GVJ_t *job, const char *s, int len)
+static size_t gv_channel_writer (GVJ_t *job, const char *s, size_t len)
 {
     return PHPWRITE(s, len);
 }
